# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: Anagrafica Radiologica API
  version: 1.0.0
  description: API per la gestione di organizzazioni, contenitori e apparecchiature radiologiche

paths:
  /api/apparecchiature:
    post:
      operationId: createApparecchiatura
      tags:
        - Apparecchiature
      summary: Crea una nuova apparecchiatura radiologica
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApparecchiaturaRequest'
      responses:
        '201':
          description: Apparecchiatura creata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApparecchiaturaResponse'
        '400':
          description: Dati non validi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Numero di serie duplicato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/organizzazioni/{id}/tree:
    get:
      operationId: getOrganizzazioneTree
      tags:
        - Organizzazioni
      summary: Recupera l'albero gerarchico completo dell'organizzazione
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Albero dell'organizzazione
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizzazioneTree'
        '404':
          description: Organizzazione non trovata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ApparecchiaturaRequest:
      type: object
      required:
        - nome
        - tipologia
        - numeroDiSerie
        - dataInstallazione
        - organizzazioneId
      properties:
        nome:
          type: string
          minLength: 1
          maxLength: 255
        tipologia:
          type: string
          enum: [TAC, RISONANZA, RX, MAMMOGRAFO, ECOGRAFO]
        numeroDiSerie:
          type: string
          minLength: 1
          maxLength: 100
        dataInstallazione:
          type: string
          format: date
        organizzazioneId:
          type: integer
          format: int64
        contenitoreId:
          type: integer
          format: int64
          nullable: true

    ApparecchiaturaResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        tipologia:
          type: string
        numeroDiSerie:
          type: string
        dataInstallazione:
          type: string
          format: date
        organizzazioneId:
          type: integer
          format: int64
        contenitoreId:
          type: integer
          format: int64

    OrganizzazioneTree:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        contenitori:
          type: array
          items:
            $ref: '#/components/schemas/ContenitoreNode'
        apparecchiature:
          type: array
          items:
            $ref: '#/components/schemas/ApparecchiaturaResponse'

    ContenitoreNode:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        sottoContenitori:
          type: array
          items:
            $ref: '#/components/schemas/ContenitoreNode'
        apparecchiature:
          type: array
          items:
            $ref: '#/components/schemas/ApparecchiaturaResponse'

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        message:
          type: string
        details:
          type: array
          items:
            type: string